pyproject.toml
[project]
name = "memory-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=0.1.0",
    "langgraph>=0.3.0",
    "mcp>=1.0.0",
    "mem0ai>=1.0.4",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
